initial communication overhaul
serialization and deserialization is now an abstract class message
types can inherit from (and every message type is now a class that
implements the serialization and deserialization methods). Message is
now a class that concerns itself with sending and receiving
messages (and consequently serializing/deserializing them) so now a
stateful conversion can be made, which permits a receive method that has a
timeout (and stores the unfinished buffer and current "location" of
received data), which was the reason for this overhaul, other than conciseness.