fixed bug where recv_for exited loop after first timeout
since it compared if receive was different from -1 in order to continue
looping, and it would be set to -1 also when recv timed out, received
being -1 caused loop be broken out from. so, to be sure, received is set
to 0 after seeing it was due to a socket timeout and the loop doesn't
depend on received being different from -1 anymore.