minor changes to authQueue cleanup and naming
I just noticed that i'm creating a god object inside master, and that
the cleanup process is generalizable to any list that can't have its
entries deleted in place, so i will need to extract it to an object so
it can be used in all lists on master or node clients. So, it's best to
do those kinds of changes now so it's easier to extract all of this
later.