generalizing managed list cleanup
since the cleanup of a list is basically deciding if it should be
cleaned up (an heuristic) and then cleaning it, the heuristic can be
implemented by the derived class and the cleanup process implemented by
the base class. The unary predicate used on the remove_if call inside
cleanList could also be customizable by the base class, but i'll think
about that later.