virtualizing deletion process and removing redundancy
sadly there is no way to make a virtual predicate since a predicate has
to be static and functions can't be static and virtual at the same time.
So, the part where the remove_if is called was virtualized instead, so
that derived classes can define the predicate that they want, with a
sane default.