implemented overseer function to manage managedList
In order to efficiently operate over a mutable size list that can have
spikes in the amount of items in it, it's best to have a function that
can start new threads to operate on it if it gets too big and stop
threads once they are unecessary. Threads get started if a heuristic
function returns a positive number (that denotes the number of new
threads to be started), and get stopped by sending the necessary amount
of stopSignals to scale back amount of running threads. It would be
interesting to have a max amount of threads that the manager list can
have so the system running it isn't thread bombed.