properly cleaning threads on threadList
i just noticed that since the class constantly starts and stops threads,
it will basically have a very big memory leak if nothing cleans the
thread list. to do that they need to be un-detached in order to try to
join/wait for each one so that the ones that are sucessfully joined can
be cleaned. Also, made basic constructors so the managedList can be set
to ran either synchronously or asynchronously.